requirements.txt
gradio>=4.0.0
langchain
langchain-core
langchain-google-genai
langgraph
pydantic
python-dotenv
pandas
llama-parse
ipykernel
langchain-community
wikipedia
langchain-openai
langgraph.checkpoint.sqlite
PyMuPDF
paddlepaddle
paddleocr
Pillow
opencv-python
